Refactor App for more OOP
There is repeated and similar code when determining input commands.

Let's create a seperate Command class handle the actions outside of App
class to imrpove readability of code.
