+++ src/main/java/InvalidCommand.java
@@ -1,0 +1,12 @@
+public class InvalidCommand extends Command {
+    protected String inst;
+
+    public InvalidCommand(String inst) {
+        this.inst = inst;
+    }
+
+    @Override
+    public void execute(Storage taskList, Ui ui) {
+        ui.invalidCommand();
+    }
+}
